<FEAT> 시간 삭제 예외처리 구현

+++ src/main/java/roomescape/service/TimeService.java
@@ -2,7 +2,9 @@ package roomescape.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import roomescape.entity.Reservation;
 import roomescape.entity.Time;
+import roomescape.exception.TimeDeleteException;
 import roomescape.repository.ReservationDao;
 import roomescape.repository.TimeDao;
 
@@ -30,6 +32,10 @@ public class TimeService {
     }
 
     public void deleteReservationTime(Long id) {
+        List<Reservation> reservations = reservationDao.findByTimeId(id);
+        if (!reservations.isEmpty()) {
+            throw new TimeDeleteException("해당 시간 id의 예약이 존재합니다");
+        }
         timeDao.delete(id);
     }
 }
